Added the main types to represent the interactions
A Command is the complete entity send by the frontend and the robot.
The cleaning pattern is the pattern the robot will clean the court
with.
The abort reason is the reason why the robot has had to stop,
therefore, all of the mid-task stopping logic needs to be contained on
the robot, for example obstacle.
Finally, the instruction is thing the robot will do.